split tests into 3 different files
